Allow configuration of NGINX Plus API access (#3066)
Allow configuration of NGINX Plus API access.

Problem: We want the ability to configure access to the NGINX Plus dashboard.

Solution: Add a field AllowedAddresses in the NginxProxy API to allow the user to configure access to the NGINX Plus dashboard.

Testing: Unit tests. Manually configured an NginxProxy resource, set up a Kind cluster with a Nodeport following our Get started guide, and confirmed that I could access the dashboard without needing to port-forward. Verified nginx plus template code is present when running on nginx plus and not present when running on nginx oss.