Remove documentation site and GitHub Action,  update README references (#3047)
As of late December, NGINX has open sourced its documentation
repository. The next stage of this effort is to consolidate all of the
existing documentation into the single repository.

This commit removes the entire `/site` folder that was being used to
manage the documentation for NGINX Gateway Fabric separately from the
main pipeline, as well as the GitHub action used for deployment.

It also updates references to the folder to instead point towards the
documentation repository, where all updates will be made in the future.