Support sending chunked files to agent (#3390)
Problem: If nginx.conf files are too large, they can't be sent over the existing gRPC stream to agent.

Solution: Agent has added an API and helper function to send files in chunks if over a certain size limit. Our controller will implement this API for agent to call when in this condition.