Add fix to route parentRef internal logic (#3418)
Internally generate ParentRefs for each listener in the Gateway if a route does not specify a sectionName.

Problem: When a route does not specify a sectionName, it attaches to all attachable listeners on a Gateway. This is a problem internally when we try to bind the route to the listeners where duplicated hostnames between listeners can create AcceptedHostname conflicts and multiple ports on the listeners can overwrite the refStatus.ListenerPort.

Solution: When a route does not specify a sectionName, generate a ParentRef for each listener in the Gateway.

* Remove duplicate parentRefs from httproute status

* Add some function comments

* Fix conformance test

* Revert parentRefIndex